Synchronize type int to Long

+++ matching-service/src/main/java/com/g13cs3219/matching_service/repositories/MatchingPool.java
@@ -8,6 +8,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ScanOptions;
 import org.springframework.data.redis.core.ZSetOperations;
 import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import com.g13cs3219.matching_service.dto.requests.JoinRequest;
 import com.g13cs3219.matching_service.dto.responses.MatchResult;
@@ -46,7 +47,7 @@ public class MatchingPool {
     /**
      * Find a match for a user based on the specified topic and difficulty.
      */
-    public Optional<MatchResult> findMatch(int userId, String topic, String difficulty, String type) {
+    public Optional<MatchResult> findMatch(Long userId, String topic, String difficulty, String type) {
         String key = buildKey(topic, difficulty);
         Optional<MatchResult> match = findExactMatch(userId, key);
         if (match.isPresent()) return match;
@@ -101,7 +102,7 @@ public class MatchingPool {
         cursor.close();
     }
 
-    private Optional<MatchResult> findExactMatch(int userId, String key) {
+    private Optional<MatchResult> findExactMatch(Long userId, String key) {
         if (key == null) return Optional.empty();
 
         Set<String> candidates = redisTemplate.opsForZSet().range(key, 0, 0);
@@ -113,14 +114,14 @@ public class MatchingPool {
         redisTemplate.opsForZSet().remove(key, match);
         return Optional.of(MatchResult.builder()
                 .userId1(userId)
-                .userId2(Integer.parseInt(match))
+                .userId2(Long.parseLong(match))
                 .topic(key.split(":")[1])
                 .difficulty(key.split(":")[2])
                 .build()
         );
     }
 
-    private Optional<MatchResult> findSameDifficultyMatch(int userId, String difficulty) {
+    private Optional<MatchResult> findSameDifficultyMatch(Long userId, String difficulty) {
         if (difficulty == null) return Optional.empty();
 
         ScanOptions options = ScanOptions.scanOptions().match("queue:*:" + difficulty).count(100).build();
